Lord, let this be the last round of review, for the files are split and my spirit also
BWG.jl (3588) -> octree.jl (1029, octant + octree + tables) and
forest.jl (2578, forest, transforms, balancing, iterator, creategrid).
ncgrid.jl and constraints.jl unchanged.

test_p4est.jl (1964) -> test_p4est_octant.jl / test_p4est_forest.jl /
test_p4est_grid.jl; the higher-level tests move to the component they
exercise (ConformityConstraint -> test_constraints.jl, facet-skeleton
InterfaceValues -> test_interfacevalues.jl, non-conforming L2 ->
test_l2_projection.jl). test_p4est_example.jl -> integration/.

Addresses the restructuring comments of @termi-official in #780.
