+++ test/test_l2_projection.jl
@@ -1,6 +1,7 @@
 # Imports for parallel (isolated) test execution:
 using LinearAlgebra
 import SHA
+import Ferrite: getrefdim
 
 # Tests a L2-projection of integration point values (to nodal values),
 # determined from the function y = 1 + x[1]^2 + (2x[2])^2
@@ -509,4 +510,44 @@ end
     test_export(subset = true)
     test_show_l2()
     test_l2proj_errorpaths()
+end
+
+# Regression test: L2 projection on a non-conforming grid must reproduce a function
+# that lies in the hanging-node constrained ansatz space exactly. This guards the
+# condensation of the mass matrix and the projection rhs in the L2Projector
+# (assembling only the upper triangle before condensation, or applying the
+# solution-style `apply!` to the rhs, both silently corrupt the projection).
+@testset "L2 projection on non-conforming grids" begin
+    @testset "$geometry" for geometry in (Quadrilateral, Hexahedron)
+        dim = getrefdim(geometry)
+        grid = generate_grid(geometry, ntuple(_ -> 2, dim))
+        forest = ForestBWG(grid, 3)
+        Ferrite.refine!(forest, [1])
+        Ferrite.balanceforest!(forest)
+        ncgrid = Ferrite.creategrid(forest)
+        # The refined grid must contain hanging nodes for this test to be meaningful
+        @test !isempty(ncgrid.conformity_info)
+
+        ip = geometric_interpolation(geometry)
+        qr = QuadratureRule{getrefshape(ip)}(2)
+        cv = CellValues(qr, ip, ip)
+
+        proj = L2Projector(ncgrid)
+        add!(proj, collect(1:getncells(ncgrid)), ip; qr_rhs = qr)
+        close!(proj)
+
+        # u(x) = x is componentwise linear, hence in the constrained ansatz space
+        qpdata = map(1:getncells(ncgrid)) do cellid
+            coords = getcoordinates(ncgrid, cellid)
+            reinit!(cv, coords)
+            [spatial_coordinate(cv, qp, coords) for qp in 1:getnquadpoints(cv)]
+        end
+        projected = project(proj, qpdata, qr)
+
+        for cell in CellIterator(proj.dh)
+            for (dof, x) in zip(celldofs(cell), getcoordinates(cell))
+                @test projected[dof] ≈ x atol = 1.0e-12
+            end
+        end
+    end
 end
